+++ src/examples/strings.test.ts
@@ -1,6 +1,6 @@
-import { Lexer } from "../../Lexer";
-import { Parser } from "../../Parser";
-import { Interpreter } from "../Interpreter";
+import { Lexer } from "../Lexer";
+import { Parser } from "../Parser";
+import { Interpreter } from "../Interpreter/Interpreter";
 
 function run(script: string): undefined | boolean | number | string {
   const tokens = new Lexer(script).tokenize();
